Fix get{Min,Max}Elements() usage
These return a nullable, UTs should be using
assertEquals(Object, Object) to compare them.

+++ yang/yang-parser-rfc7950/src/test/java/org/opendaylight/yangtools/yang/stmt/EffectiveUsesRefineAndConstraintsTest.java
@@ -106,8 +106,8 @@ public class EffectiveUsesRefineAndConstraintsTest {
         assertFalse(listSchemaNode.isConfiguration());
 
         ElementCountConstraint listConstraints = listSchemaNode.getElementCountConstraint().get();
-        assertEquals(10, listConstraints.getMinElements().intValue());
-        assertEquals(20, listConstraints.getMaxElements().intValue());
+        assertEquals(10, listConstraints.getMinElements());
+        assertEquals(20, listConstraints.getMaxElements());
         assertEquals(1, listSchemaNode.getMustConstraints().size());
     }
 
@@ -141,8 +141,8 @@ public class EffectiveUsesRefineAndConstraintsTest {
         assertTrue(listSchemaNode.isConfiguration());
 
         ElementCountConstraint listConstraints = listSchemaNode.getElementCountConstraint().get();
-        assertEquals(5, listConstraints.getMinElements().intValue());
-        assertEquals(7, listConstraints.getMaxElements().intValue());
+        assertEquals(5, listConstraints.getMinElements());
+        assertEquals(7, listConstraints.getMaxElements());
         assertEquals(2, listSchemaNode.getMustConstraints().size());
     }
 }
